requirements.txt
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
psycopg2==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
setuptools==69.5.1
social-auth-app-django==5.4.1
drf-yasg==1.21.7
redis==5.0.4
celery==5.4.0